Modified flag usage. Split RSA related crypto into seperate files
