+++ commands/author.go
@@ -13,7 +13,7 @@ func init() {
 var (
 	authorCmd = &cobra.Command{
 		Use:   "authors",
-		Short: "Display the list of Foil authors",
+		Short: "Display the list of foil authors",
 		Long:  ``,
 		Run: func(cmd *cobra.Command, args []string) {
 			fmt.Println("Authors: Brian Vohaska <bvohaska@gmail.com>")
