+++ commands/foil.go
@@ -7,11 +7,11 @@ import (
 func init() {
 	// List of all top-level flags used in Foil
 
-	FoilCmd.PersistentFlags().StringVarP(&inputPath, "in", "", "", "Read input as file from PATH=[string]")
-	FoilCmd.PersistentFlags().StringVarP(&outputPath, "out", "", "", "Save output as file located at PATH=[string]")
-	FoilCmd.PersistentFlags().StringVarP(&stdInString, "textin", "", "", "Read input from StdIn as [string]")
-	FoilCmd.PersistentFlags().BoolVarP(&Verbose, "verbose", "v", false, "Display verbose output")
-	FoilCmd.PersistentFlags().BoolVarP(&stdOutBool, "textout", "", false, "Display output on StdOut")
+	FoilCmd.PersistentFlags().StringVarP(&inputPath, "in", "", "", "read input as file from PATH=[string]")
+	FoilCmd.PersistentFlags().StringVarP(&outputPath, "out", "", "", "save output as file located at PATH=[string]")
+	FoilCmd.PersistentFlags().StringVarP(&stdInString, "textin", "", "", "read input from StdIn as [string]")
+	FoilCmd.PersistentFlags().BoolVarP(&Verbose, "verbose", "v", false, "display verbose output")
+	FoilCmd.PersistentFlags().BoolVarP(&stdOutBool, "textout", "", false, "display output on StdOut")
 
 	// Add static  commands
 	FoilCmd.AddCommand(versionCmd)
