+++ cryptospecials/cryptospecials_test.go
@@ -4,6 +4,7 @@ import (
 	"crypto/elliptic"
 	"crypto/rsa"
 	"crypto/sha256"
+	"math/big"
 	"testing"
 )
 
@@ -186,13 +187,21 @@ func TestRSAVRFGenVerify(t *testing.T) {
 
 func TestHash2Curve(t *testing.T) {
 
+	var (
+		x *big.Int
+		y *big.Int
+	)
+
 	//data := []byte("I'm a string!")
 	data := []byte("I'm a string!")
-	h := sha256.New()
+	hash256 := sha256.New()
 	ec := elliptic.P256()
 
-	err := Hash2curve(data, h, ec.Params(), 1, false)
+	x, y, err := Hash2curve(data, hash256, ec.Params(), 1, false)
 	if err != nil {
 		t.Errorf("FAIL: %v\n", err)
+	}
+	if x == zero || y == zero {
+		t.Errorf("FAIL: Zero values returned as points\n")
 	}
 }
